refactor: break the api/routes into files apart
